Fix Oracle MCP integration and improve tool response handling
- Fix JSON serialization errors for tool responses in chat.py
- Add proper content extraction for LlamaStack TextContentItem objects
- Implement unified tool response processing for both Regular and ReAct agents
- Add asyncio exception handler to suppress harmless TCP transport cleanup errors
- Update frontend LlamaStackAdapter to filter tool invocation syntax and process tool_result responses
- Update Oracle MCP server with FastMCP implementation for better LlamaStack compatibility
- Improve Oracle MCP setup documentation with complete step-by-step instructions
- Add comprehensive security validation for custom SQL queries
- Update LlamaStack configuration to include Oracle MCP server endpoint
