+++ docs/oracle-mcp-setup.md
@@ -1,44 +1,119 @@
 # Oracle MCP Server Setup Guide
 
-This guide shows how to register the Oracle MCP server with the AI Virtual Agent platform using the existing MCP server management interface.
+This guide shows how to set up the Oracle MCP server with Oracle 23ai database and register it with the AI Virtual Agent platform.
 
 ## Overview
 
-The Oracle MCP server provides 9 business analytics tools for TPC-DS data analysis:
+The Oracle MCP server provides business analytics tools for TPC-DS data analysis:
 
-- **health_check** - Server health monitoring
-- **get_tpcds_summary** - TPC-DS tables and row counts  
-- **get_customer_insights** - Customer demographic analysis
-- **get_sales_analytics** - Sales transaction analytics
-- **get_top_selling_products** - Product performance metrics
-- **get_inventory_insights** - Inventory analysis with alerts
-- **get_store_performance** - Store performance comparison
-- **execute_custom_query** - Safe custom SQL execution
-- **get_kpi_dashboard** - Key performance indicators
+- **health_check** - Server health and database connectivity monitoring
+- **get_tpcds_summary** - TPC-DS tables overview and row counts  
+- **get_customer_insights** - Customer demographic analysis with configurable limits
+- **get_sales_analytics** - Sales transaction analytics with date range filtering
+- **execute_custom_query** - Safe custom SQL execution with security constraints
 
 ## Prerequisites
 
-1. **Oracle Database**: Oracle 23ai with TPC-DS data loaded
-2. **Oracle MCP Server**: Deployed and running on port 8003
-3. **Database Connectivity**: Oracle MCP server can connect to Oracle DB
-4. **Admin Access**: Admin role in AI Virtual Agent platform
+1. **Oracle Database**: Oracle 23ai with TPC-DS data loaded and accessible
+2. **Database Connection**: Port forwarding or direct connectivity to Oracle database
+3. **Oracle Credentials**: Valid system user credentials for Oracle database
+4. **Python Environment**: Python 3.8+ with required dependencies
+5. **Admin Access**: Admin role in AI Virtual Agent platform for MCP server registration
 
-## Registration Steps
+## Step 1: Setup Oracle Database Connection
+
+### Option A: Local Development with Port Forwarding
+
+If you have Oracle 23ai running in Kubernetes/OpenShift:
+
+```bash
+# Get the Oracle pod name
+oc get pods -n an-oracle-23ai
+
+# Start port forwarding (keep this running in a separate terminal)
+oc port-forward -n an-oracle-23ai oracle23ai-0 1521:1521
+```
+
+### Option B: Direct Database Connection
+
+Update the connection details in the Oracle MCP server configuration to point directly to your Oracle database.
+
+## Step 2: Setup Oracle MCP Server
+
+### Get Oracle Database Credentials
+
+First, get the correct Oracle password from your Kubernetes deployment:
+
+```bash
+# Get the Oracle password from the secret
+oc get secret oracle23ai -n an-oracle-23ai -o jsonpath='{.data.password}' | base64 -d
+```
+
+### Configure Oracle MCP Server
+
+1. Navigate to the Oracle MCP server directory:
+```bash
+cd mcpservers/mcp_oracle
+```
+
+2. Create the `.env` file with correct Oracle credentials:
+```bash
+# Oracle Database Configuration
+ORACLE_USER=SYSTEM
+ORACLE_PASSWORD=<password_from_secret_above>
+ORACLE_HOST=localhost
+ORACLE_PORT=1521
+ORACLE_SERVICE_NAME=freepdb1
+
+# MCP Server Configuration  
+MCP_SERVER_PORT=8003
+MCP_SERVER_HOST=0.0.0.0
+
+# Security Settings
+ENVIRONMENT=development
+
+# Logging Configuration
+LOG_LEVEL=INFO
+```
+
+3. Install dependencies and start the Oracle MCP server:
+```bash
+# Install Python dependencies
+pip install -r requirements.txt
+
+# Start the Oracle MCP server (keep this running)
+python oracle_fastmcp_server.py
+```
+
+The server should start and show:
+```
+INFO:     Uvicorn running on http://0.0.0.0:8003 (Press CTRL+C to quit)
+```
+
+### Verify Oracle MCP Server
+
+Test the server is working:
+```bash
+# Test basic connectivity
+curl http://localhost:8003/
+
+# Test health check (should return Oracle database status)
+curl -X POST http://localhost:8003/call \
+  -H "Content-Type: application/json" \
+  -d '{"method": "health_check", "params": {}}'
+```
+
+## Step 3: Register with AI Virtual Agent Platform
 
 ### 1. Access MCP Servers Configuration
 
-1. Log into the AI Virtual Agent platform
+1. Log into the AI Virtual Agent platform at http://localhost:5173
 2. Navigate to **Configuration > MCP Servers** (admin only)
 3. Click the **+** icon to add a new MCP server
 
 ### 2. Fill Out Registration Form
 
 Use these exact values in the MCP server registration form:
-
-**Toolgroup ID:**
-```
-mcp-oracle
-```
 
 **Name:**
 ```
@@ -52,19 +127,26 @@ TPC-DS business analytics tools for Oracle Database
 
 **Endpoint URL:**
 ```
-http://oracle-mcp-service:8003
+http://192.168.2.199:8003/sse
+```
+
+**Tool Group ID:**
+```
+mcp::oracle_analytics
 ```
 
 **Configuration (JSON):**
 ```json
 {
-  "oracle_host": "oracle-service.an-oracle-23ai.svc.cluster.local",
-  "oracle_port": "1521",
-  "service_name": "FREEPDB1",
-  "username": "tpcds",
-  "environment": "development"
+  "name": "Oracle Analytics",
+  "description": "TPC-DS business analytics tools for Oracle Database"
 }
 ```
+
+> **Important Notes:**
+> - Use the exact IP address and `/sse` endpoint path as shown above
+> - The Tool Group ID must include the `mcp::` prefix
+> - The endpoint URL should match your local network IP address (check with `ifconfig` or `ip addr`)
 
 ### 3. Submit Registration
