+++ frontend/src/adapters/llamaStackAdapter.ts
@@ -27,6 +27,11 @@ export const LlamaStackParser: LlamaStackParserType = {
 
       // Handle text content which should be shown to the user
       if (json.type === 'text' && json.content) {
+        // Filter out tool invocation brackets like [tool_name()] that come from LLM responses
+        if (json.content.match(/^\[[\w_]+\([^[\]]*\)\]$/)) {
+          return null; // Skip displaying tool invocation syntax
+        }
+        
         if (agentType === 'ReAct') {
           const result = processStreamingReActResponse(json.content);
           return result;
@@ -35,9 +40,42 @@ export const LlamaStackParser: LlamaStackParserType = {
         return json.content;
       }
 
-      // Handle tool use - convert to visual indication that tool is being used
+      // Handle tool use - skip during streaming since tool_result will show the actual output
       if (json.type === 'tool' && json.content) {
-        return `[Tool: ${json.tool?.name || 'unknown'}] ${json.content}\n`;
+        return null; // Skip tool invocation messages during streaming
+      }
+
+      // Handle tool results - process them like stored messages for consistency
+      if (json.type === 'tool_result' && json.content) {
+        // Extract clean content, skipping raw JSON formatting
+        let cleanContent = json.content;
+        
+        // Remove the "**tool_name result:**" header if present
+        cleanContent = cleanContent.replace(/^\*\*[^*]+\*\*\s*\n?/, '');
+        
+        // Remove JSON code block wrappers if present
+        cleanContent = cleanContent.replace(/^```json\s*\n?/, '').replace(/\n?```$/, '');
+        
+        // Skip if content looks like raw TextContentItem format
+        if (cleanContent.includes('TextContentItem(text=')) {
+          return null; // Skip showing raw TextContentItem responses
+        }
+        
+        // Try to parse and extract meaningful content
+        try {
+          const parsed = JSON.parse(cleanContent);
+          // If it's a tool response with meaningful data, format it nicely
+          if (parsed && typeof parsed === 'object') {
+            return JSON.stringify(parsed, null, 2);
+          }
+        } catch (e) {
+          // If not JSON, return as-is if it's meaningful text
+          if (cleanContent.trim() && !cleanContent.includes('TextContentItem')) {
+            return cleanContent;
+          }
+        }
+        
+        return null; // Skip malformed or empty responses
       }
 
       // Handle reasoning (thought process)
